+++ components/EditPanel/editPanel.module.scss
@@ -2,20 +2,4 @@
   display: flex;
   flex-direction: row;
   gap: 8px;
-  .button {
-    width: 45px;
-    height: 45px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 24px;
-    border-radius: 8px;
-    background: var(--lt-plates-background);
-    cursor: pointer;
-    transition: .2s background;
-    border: none;
-    &.active, &:hover {
-      background: var(--lt-hovered-plates-background);
-    }
-  }
 }
