+++ components/SideLinks/sideLinks.module.scss
@@ -1,0 +1,3 @@
+.container {
+  padding-left: 8px;
+}
