--- components/Sidebar/Sidebar.js
@@ -1,38 +1,0 @@
-import classes from './sidebar.module.scss'
-import Link from "../Link/Link";
-import {Github, Telegram, Vk, WideWorldNet} from "../SvgIcons/SvgIcons";
-import Layout from "../Layout/Layout";
-
-export default function Sidebar() {
-    return (
-        <aside className={classes.container}>
-            <Layout.Container>
-                <Link href={'/about_project'}>about project</Link>
-                <Link href={'/support'}>support</Link>
-            </Layout.Container>
-            <Layout.Container>
-                <Link.Button href={'https://github.com/Revilise'}>
-                    <Github/>
-                    project's git
-                </Link.Button>
-                <SocialLink url={"https://vk.com/Revilise"} socialName={"vk"} Img={Vk} />
-                <SocialLink url={"https://t.me/Revilise"} socialName={"Telegram"} Img={Telegram} />
-                <SocialLink url={"https://portf.com/Revilise"} socialName={"Site"} Img={WideWorldNet} />
-            </Layout.Container>
-        </aside>
-    )
-}
-
-function SocialLink({ Img, socialName, url }) {
-    return (
-        <Link href={url}>
-            <div className={classes.social_link}>
-                <Img/>
-                <div>
-                    <h3 className={classes.title}>{socialName}</h3>
-                    <span className={classes.link}>{url.replace('https://', '')}</span>
-                </div>
-            </div>
-        </Link>
-    )
-}
