--- components/Sidebar/sidebar.module.scss
@@ -1,19 +1,0 @@
-.container {
-  gap: 24px;
-  grid-area: sidebar;
-  display: flex;
-  flex-direction: column;
-}
-
-.social_link {
-  display: flex;
-  align-items: center;
-  padding: 12px 8px 12px 16px;
-  gap: 24px;
-  .title {
-
-  }
-  .link {
-    font-size: 14px;
-  }
-}
