+++ pages/index.js
@@ -3,7 +3,6 @@ import PostsList from "../components/PostsList/PostsList";
 import postsController from "../db/controllers/posts.controller";
 import Header from "../components/Header/Header";
 import Link from "../components/Link/Link";
-import Sidebar from "../components/Sidebar/Sidebar";
 import Main from "../components/Main/Main";
 import {useEffect, useState} from "react";
 
@@ -17,10 +16,9 @@ export default function Home() {
     return (
         <Layout title={"home"}>
             <Header search>
-                <Link.Button href={'/create_post'}>create new post</Link.Button>
+                <Link.Button href={'/create_post'}>create</Link.Button>
             </Header>
             <Main>
-                <Main.Aside children={<Sidebar />}/>
                 <Main.Section children={<PostsList posts={posts} />} />
             </Main>
         </Layout>
