+++ pages/post/[url].js
@@ -5,7 +5,6 @@ import {useRouter} from "next/router";
 import {useEffect, useState} from "react";
 import Loading from "../../components/Loading/Loading";
 import Header from "../../components/Header/Header";
-import Sidebar from "../../components/Sidebar/Sidebar";
 import Main from "../../components/Main/Main";
 import Search from "../../components/Search/Search";
 import Button from "../../components/Button/Button";
@@ -33,12 +32,9 @@ export default function Post() {
     return (
         <Layout title={"Sog"}>
             <Header search>
-                <Link.Button href={'/create_post'}>create new post</Link.Button>
+                <Link.Button href={'/create_post'}>create</Link.Button>
             </Header>
             <Main>
-                <Main.Aside>
-                    <Sidebar/>
-                </Main.Aside>
                 <Main.Section>
                     { isFetch ? <Loading /> : <PostDetails {...data} deletePost={deletePost}/>}
                 </Main.Section>
