+++ pages/search/[search].js
@@ -2,27 +2,26 @@ import Layout from "../../components/Layout/Layout";
 import Header from "../../components/Header/Header";
 import Link from '../../components/Link/Link'
 import Main from "../../components/Main/Main";
-import Sidebar from "../../components/Sidebar/Sidebar";
 import PostsList from "../../components/PostsList/PostsList";
 import {useEffect, useState} from "react";
 import {useRouter} from "next/router";
-import Search from "../../components/Search/Search";
 import postsController from "../../db/controllers/posts.controller";
 
 export default function _Search() {
     const [posts, setPosts] = useState([]);
     const router = useRouter();
+    const search = router.query.search;
     useEffect(() => {
-        postsController.getByString(router.query.search).then(res => setPosts(res));
-    }, [router.query.search])
+        if (search)
+            postsController.getByString(router.query.search).then(res => setPosts(res));
+    }, [search])
 
     return (
         <Layout title={"search..."}>
-            <Header search>
-                <Link.Button href={'/create_post'}>create new post</Link.Button>
+            <Header>
+                <Link.Button href={'/create_post'}>create</Link.Button>
             </Header>
             <Main>
-                <Main.Aside children={<Sidebar />} />
                 <Main.Section>
                     <PostsList posts={posts} />
                 </Main.Section>
