+++ resorses/draftResourses.js
@@ -1,0 +1,13 @@
+import Draft from "draft-js";
+
+export const emptyContentState = Draft.convertFromRaw({
+    entityMap: {},
+    blocks: [
+        {
+            text: '',
+            key: 'foo',
+            type: 'unstyled',
+            entityRanges: [],
+        },
+    ],
+});
