+++ resorses/resourses.js
@@ -1,0 +1,6 @@
+
+
+export const createPostSideLinks = [
+    {href: "/", title: "instruction"}
+]
+
